# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tresleai Rest API ",
    "description": "Tresleai Facade Microservice API",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1.1/admin/apps/{app_name}": {
      "get": {
        "tags": [
          "crate"
        ],
        "summary": "GET handler to get an app.",
        "operationId": "get_app",
        "parameters": [
          {
            "name": "app_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App retrieved succesfully."
          },
          "400": {
            "description": "Invalid Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Tresle error occurred. Please save reference id: {} and contact support."
          }
        }
      },
      "delete": {
        "tags": [
          "crate"
        ],
        "summary": "DELETE handler to delete an app and other associated resources.",
        "operationId": "delete_app",
        "parameters": [
          {
            "name": "app_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App deleted succesfully."
          },
          "400": {
            "description": "Invalid Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Tresle error occurred. Please save reference id: {} and contact support."
          }
        }
      }
    },
...
    "/api/v1.1/admin/nodes/errors/{app_name}": {
      "get": {
        "tags": [
          "crate"
        ],
        "summary": "GET handler to fetch errors while processing/extracting knowledge nodes for an app between two timestamps.",
        "operationId": "get_knowledge_nodes_errors_handler",
        "parameters": [
          {
            "name": "start_timestamp",
            "in": "query",
            "description": "start timestamp.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_timestamp",
            "in": "query",
            "description": "end timestamp.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "page number.",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "page limit.",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "app_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Errors while processing knowledge nodes for app fetched successfully."
          },
          "400": {
            "description": "Invalid Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Tresle error occurred. Please save reference id: {} and contact support."
          }
        }
      }
    },
    "/api/v1.1/admin/nodes/{app_name}": {
      "get": {
        "tags": [
          "crate"
        ],
        "summary": "GET handler to fetch knowledge nodes for an app between two timestamps.",
        "operationId": "get_knowledge_nodes_handler",
        "parameters": [
          {
            "name": "start_timestamp",
            "in": "query",
            "description": "start timestamp.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_timestamp",
            "in": "query",
            "description": "end timestamp.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "knowledge_node_type",
            "in": "query",
            "description": "knowledge node type.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "page number.",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "page limit.",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "app_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Knowledge nodes for app fetched successfully."
          },
          "400": {
            "description": "Invalid Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Tresle error occurred. Please save reference id: {} and contact support."
          }
        }
      }
    },
...
    "/api/v1.1/admin/search/apps/{app_name}": {
      "patch": {
        "tags": [
          "crate"
        ],
        "summary": "PATCH handler to update the search_enabled flag of an app.",
        "operationId": "update_search_enabled_handler",
        "parameters": [
          {
            "name": "search_enabled",
            "in": "query",
            "description": "search enabled flag.",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "app_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search_enabled flag updated successfully."
          },
          "400": {
            "description": "Invalid Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Tresle error occurred. Please save reference id: {} and contact support."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  }
}